Add authorize to methods

+++ Requests/NoteRequests.cs
@@ -1,6 +1,10 @@
+using System.Security.Claims;
 using FluentValidation;
+using System.Collections.Generic;
 using NoteMiniApi.Models;
 using NoteMiniApi.Services;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 
 namespace NoteMiniApi.Requests
 {
@@ -13,6 +17,10 @@ namespace NoteMiniApi.Requests
 
             app.MapGet("/notes/{id}", NoteRequests.GetNoteById)
             .Produces<Note>()
+            .Produces(StatusCodes.Status404NotFound);
+
+            app.MapGet("/notes/mynotes", NoteRequests.GetUserNotes)
+            .Produces<List<Note>>()
             .Produces(StatusCodes.Status404NotFound);
 
 
@@ -51,35 +59,59 @@ namespace NoteMiniApi.Requests
             }
             return Results.Ok(note);
         }
-
-        public static IResult CreateNote(INoteService service, Note note)
+        [Authorize]
+        public static IResult CreateNote(INoteService service, Note note, ClaimsPrincipal User)
         {
+            var userId = (User.FindFirst(ClaimTypes.NameIdentifier).Value);
+            service.AddNote(note, userId);
 
-            service.AddNote(note);
-
-            return Results.Created($"/notes/{note.Id}", note);
+           // return Results.Created($"/notes/{note.Id}", note);
+           return Results.Ok();
         }
-
-        public static IResult DeleteNote(INoteService service, int Id)
+        [Authorize]
+        public static IResult DeleteNote(INoteService service, int Id, ClaimsPrincipal User)
         {
+            
+            var userId = (User.FindFirst(ClaimTypes.NameIdentifier).Value);
+
             var note = service.GetNoteById(Id);
             if(note == null)
             {
                 return Results.NotFound();
             }
+            if(!service.isUserNote(userId, Id))
+            {
+                return Results.Forbid();
+            }
             service.DeleteNote(note);
             return Results.NoContent();
         }
-
-        public static IResult UpdateNote(INoteService service, Note note)
+        [Authorize]
+        public static IResult UpdateNote(INoteService service, Note note, ClaimsPrincipal User)
         {
+            var userId = (User.FindFirst(ClaimTypes.NameIdentifier).Value);
             if(note == null)
             {
                 return Results.NotFound();
             }
+             if(!service.isUserNote(userId, note.Id))
+            {
+                return Results.Forbid();
+            }
             note.CreateAt = System.DateTime.Now;
             service.UpdateNote(note);
             return Results.NoContent();
+        }
+        [Authorize]
+        public static IResult GetUserNotes(INoteService service, ClaimsPrincipal User)
+        {
+            var userId = (User.FindFirst(ClaimTypes.NameIdentifier).Value);
+            var notes = service.GetUserNotes(userId);
+            if(notes == null)
+            {
+                return Results.NotFound();
+            }
+            return Results.Ok(notes);
         }
     }
 }
